+++ source/3_games/pokemon/include/public/pokemon/render_constants.h
@@ -22,11 +22,11 @@ namespace pokemon
 
 		constexpr u8 g_square_width_px = g_num_tiles_per_square_row * g_tile_width_px;
 		constexpr u8 g_square_height_px = g_num_tiles_per_square_column * g_tile_height_px;
-		constexpr u8 g_square_size_px = g_square_width_px * g_square_height_px;
+		constexpr u16 g_square_size_px = g_square_width_px * g_square_height_px;
 
 		constexpr u8 g_block_width_px = g_num_tiles_per_block_row * g_tile_width_px;
 		constexpr u8 g_block_height_px = g_num_tiles_per_block_column * g_tile_height_px;
-		constexpr u8 g_block_size_px = g_block_width_px * g_block_height_px;
+		constexpr u16 g_block_size_px = g_block_width_px * g_block_height_px;
 
 		constexpr u8 g_screen_width = 160;
 		constexpr u8 g_screen_height = 144;
